+++ Utils/MainInstances.gd
@@ -2,3 +2,4 @@ extends Resource
 
 var songFilePath = null
 var beatDetector = null
+var player = null